+++ examples/opentelemetry-web/examples/xml-http-request/index.js
@@ -4,6 +4,7 @@ const { WebTracerProvider } = require( '@opentelemetry/sdk-trace-web');
 const { XMLHttpRequestInstrumentation } = require( '@opentelemetry/instrumentation-xml-http-request');
 const { ZoneContextManager } = require( '@opentelemetry/context-zone');
 const { OTLPTraceExporter } = require( '@opentelemetry/exporter-trace-otlp-http');
+const { CompressionAlgorithm } = require('@opentelemetry/otlp-exporter-base');
 const { B3Propagator } = require( '@opentelemetry/propagator-b3');
 const { registerInstrumentations } = require( '@opentelemetry/instrumentation');
 
@@ -13,7 +14,11 @@ const providerWithZone = new WebTracerProvider();
 // to your exporter. Using the SimpleSpanProcessor here as it sends the spans immediately to the
 // exporter without delay
 providerWithZone.addSpanProcessor(new SimpleSpanProcessor(new ConsoleSpanExporter()));
-providerWithZone.addSpanProcessor(new SimpleSpanProcessor(new OTLPTraceExporter()));
+
+const exporterOptions = {};
+// Note: to use GZIP compression during export, use the following exporterOptions
+// const exporterOptions = { compression: CompressionAlgorithm.GZIP, headers:  { 'force-xhr': 'custom header is required to force xhr use' } };
+providerWithZone.addSpanProcessor(new SimpleSpanProcessor(new OTLPTraceExporter( exporterOptions)));
 
 providerWithZone.register({
   contextManager: new ZoneContextManager(),