+++ experimental/packages/otlp-exporter-base/src/platform/node/index.ts
@@ -16,4 +16,5 @@
 
 export { OTLPExporterNodeBase } from './OTLPExporterNodeBase';
 export { sendWithHttp, createHttpAgent, configureCompression } from './util';
-export { OTLPExporterNodeConfigBase, CompressionAlgorithm } from './types';
+export { OTLPExporterNodeConfigBase } from './types';
+export { CompressionAlgorithm } from '../../types';